Switched to WinAPI for process creation, and other small changes.
The main highlight of this commit is the switch to WinAPI for creating processes,
as opposed to Rust's standard built-in std::process::Command, which fixes the various
process launching issues, especially with cmd/pwsh, and also allowing for more fine
grain control over how processes are launched, with the possibility of adding new
JSON keys in the future.

Also improved the code structure, renamed variables, and just made the codebase
more consistent with itself.